backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
timm==1.0.9
albumentations==1.4.14
opencv-python-headless==4.10.0.84
huggingface_hub==0.24.6
numpy==1.26.4
